Use analysis config file hashmap in dynamo, move DynamoDBAccessTest to unit folder
