+++ src/test/java/cloud/filibuster/functional/java/basic/GrpcMockTest.java
@@ -25,10 +25,8 @@ import java.util.UUID;
 import static cloud.filibuster.instrumentation.helpers.Property.setTestAvoidRedundantInjectionsProperty;
 import static cloud.filibuster.integration.instrumentation.TestHelper.startAPIServerAndWaitUntilAvailable;
 import static cloud.filibuster.integration.instrumentation.TestHelper.stopAPIServerAndWaitUntilUnavailable;
-import static cloud.filibuster.junit.Assertions.wasFaultInjected;
 import static cloud.filibuster.junit.Assertions.wasFaultInjectedOnMethod;
 import static org.grpcmock.GrpcMock.stubFor;
-import static org.grpcmock.GrpcMock.times;
 import static org.grpcmock.GrpcMock.never;
 import static org.grpcmock.GrpcMock.unaryMethod;
 import static org.grpcmock.GrpcMock.verifyThat;
